requirements.txt
yagmail==0.15.277
schedule==1.1.0
python-dotenv==0.19.2
pathlib==1.0.1